FWA-13 - Angular code refactoring - cleanup routing
