+++ src/app/pages/home/about/components/about-banner/about-banner.component.html
