+++ src/app/pages/home/components/banner-bottom/banner-bottom.component.scss
