+++ src/app/pages/home/components/banner/banner.component.ts
