+++ src/app/pages/home/components/footer/footer.component.scss
