+++ src/app/pages/home/components/footer/footer.component.spec.ts
