+++ src/app/pages/home/components/footer/footer.component.ts
