+++ src/app/pages/home/components/section-header/section-header.component.scss
