+++ src/app/pages/home/landing/components/updates/updates.component.scss
