+++ src/app/pages/home/landing/components/view-post/view-post.component.html
